requirements.txt
django~=6.0.3
django-environ~=0.13.0
django-filter~=25.2
django-model-utils~=5.0.0
django-phonenumber-field[phonenumbers]~=8.4.0
django-stubs-ext~=6.0.3
djangorestframework~=3.17.1
pyyaml~=6.0.3
